pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roma-agent"
version = "0.1.0"
description = "Idea-to-article multi-platform blogging pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.1",
  "pydantic>=2.7.0",
  "rich>=13.7.1",
  "agent-framework>=1.0.0rc0",
  "requests>=2.32.0",
  "azure-identity>=1.17.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
roma-agent = "roma_agent.cli:main"

